fix: clear mypy --strict debt in production code; scope CI mypy to it
CI's mypy job ran to completion for the first time ever on Day 13 (earlier
runs never got past pip install) and surfaced ~258 pre-existing errors
across production code and tests/ -- debt that predates this session, not
caused by the torch/pip fixes.

Per decision: keep --strict on the actual package (agents/, api/,
retrieval/, monitoring/, evaluation/, schema/, mcp/) and fix its ~44 real
errors; scope tests/ and the root-level run_*.py/verify_phase_c*.py
one-off scripts out of mypy in pyproject.toml (same 'not part of the
AGENTS.md package layout' reasoning as the existing run_*.py ruff
per-file-ignore) rather than silently weakening strict on code that ships.

Fixes, by category:
- anthropic Messages.create calls (qa_agent, text_to_sql_agent,
  citation_judge, relevance_labeller): type tools/messages/tool_choice
  with anthropic.types instead of bare dict/list literals, which mypy's
  overload resolution can't structurally match
- langgraph add_node calls (qa_agent, retrieval_agent, orchestrator):
  scoped  -- mypy can't unify NodeInputT through
  add_node's Union-of-Protocols overloads for a plain
  Callable[[State], State]; each node's own signature is correct
  (known langgraph/mypy stub limitation)
- sys.stdout.reconfigure (eda_agent, benchmark_runner, ragas_suite,
  relevance_labeller): isinstance-guard on io.TextIOWrapper instead of
  assuming TextIO has .reconfigure
- write_jsonl: list[BaseModel] -> Sequence[BaseModel] param (list is
  invariant, Sequence isn't) so list[BenchmarkQuery]/list[RelevanceLabel]
  callers type-check
- drift_detector: np.ndarray -> npt.NDArray[np.float64] throughout
- schema/models.py: Column[list[str]] annotation for the ARRAY(String)
  column mypy couldn't infer; Engine/get_engine return types
  get_session_factory param type
- dense_index: cast() the narrowed QdrantSearchClient Protocol at
  DenseIndex.connect() (real QdrantClient is structurally wider than the
  Protocol tests inject); cast() the sentence-transformers .encode()
  return; cast(str, ...) chunk_id from payload.get() *without* a
  default -- an earlier draft used .get(chunk_id, ) here and broke
  test_missing_payload_fields_default_to_none_and_empty_text, since this
  project's Candidate dataclass isn't runtime-validated and downstream
  code relies on a missing chunk_id staying None, not becoming ""
- reranker_cohere/biencoder_eval/populate_qdrant/mcp_airflow/eda_agent:
  explicit return-type annotations/casts where a stub-untyped call
  (cohere, numpy sum, pandasai's LLM base, requests .json()) was flowing
  Any into a typed return

Verified: ruff clean, mypy --strict clean (59 source files), full pytest
suite green (429/429) -- the dense_index chunk_id regression above was
caught by the existing test suite before this commit, not after.

+++ retrieval/biencoder_eval.py
@@ -165,7 +165,7 @@ QUERIES: list[EvalQuery] = [
 
 
 def _dcg(gains: list[int]) -> float:
-    return sum((2**g - 1) / np.log2(i + 2) for i, g in enumerate(gains))
+    return float(sum((2**g - 1) / np.log2(i + 2) for i, g in enumerate(gains)))
 
 
 def ndcg_at_k(ranked_chunk_ids: list[str], judgments: dict[str, int], k: int = TOP_K) -> float:
